+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New functionality that are backwards-compatible:
 * [OLMIS-7197](https://openlmis.atlassian.net/browse/OLMIS-7197): Refreshed pending offline events indicator after saving a new stock event.
 * [OLMIS-7198](https://openlmis.atlassian.net/browse/OLMIS-7198): Added the possibility to redirect to Pending Operations screen from pending offline events link.
 * [OLMIS-7181](https://openlmis.atlassian.net/browse/OLMIS-7181): Made Stock Card Summaries cache user-specific.
+* [OLMIS-7205](https://openlmis.atlassian.net/browse/OLMIS-7205): Synchronized offline events after logging and switching from offline to online.
 
 Bug fixes:
 * [OLMIS-7220](https://openlmis.atlassian.net/browse/OLMIS-7220): Added call to facilityProgramCacheService in the routes.